엔티티 컴포넌트 수정 완료 / 테스트 필요 + 테스트 후 이전 클래스 제거

+++ src/main/java/net/biryeongtrain06/qf_stat_mod/utils/enums/Elements.java
@@ -101,7 +101,7 @@ public enum Elements {
         return element.orElse(PHYSICAL);
     }
 
-    public static float calculateDamageReduce(Elements e, int resistance, float amount) {
+    public static float calculateDamageReduce(Elements e, float resistance, float amount) {
         if (resistance == 0) {
             return amount;
         }
